Corrected addContact & removeContact

+++ controllers/contacts-controller.js
@@ -25,7 +25,7 @@ const getContactById = async (req, res) => {
 
 const addContact = async (req, res) => {
   const { _id: owner } = req.user;
-  const result = await Contact.create(...req.body, owner);
+  const result = await Contact.create({ ...req.body, owner });
   res.status(201).json(result);
 };
 
@@ -52,7 +52,7 @@ const updateFavorite = async (req, res) => {
 const removeContact = async (req, res) => {
   const { id } = req.params;
   const { _id: owner } = req.user;
-  const result = await Contact.findOneAndDelete({ _id: id });
+  const result = await Contact.findOneAndDelete({ _id: id, owner });
   if (!result) {
     throw HttpError(404, `Contact with id ${id} not found`);
   }
